requirements.txt
adb==1.3.0
lxml==4.4.1
prompt_toolkit==1.0.14
pychalk==2.0.1
PyInquirer==1.0.3
pyserial==3.4
